src/requirements.txt
Django==2.0.1
django-money==0.12.3
djangorestframework==3.7.3
olefile==0.44
psycopg2==2.7.3.2
py-moneyed==0.7.0
pytz==2017.3
gunicorn==19.8.1
django-allauth==0.36.0
django-reversion==2.0.13

